refactor(sboms): move CBOM/dup-error helpers to utils.py (#1069)
Address Copilot re-review: _is_cbom / _is_crypto_component / _is_duplicate_integrity_error
(plus the unique-constraint name) move from apis.py to the neutral sboms/utils.py, so the
backfill management command no longer imports the API/ninja layer to reach them. apis.py
imports them back; behavior unchanged.
